+++ src/engines/hfbi/bn/mod.rs
@@ -42,8 +42,8 @@ mod bn_private_tests {
     fn create_dummy_record(peso: f32) -> RecordHFBI {
         RecordHFBI {
             specie: SpecieHFBI {
-                nome_comune: "Dummy",
-                codice_specie: "DM",
+                nome_comune: "Dummy".to_string(),
+                codice_specie: "DM".to_string(),
                 autoctono: true,
                 gruppo_eco: GruppoEcoHFBI::ResidentiDiEstuario, // This field is not used by calc_bn
                 gruppo_trofico: GruppoTrofHFBI {