+++ src/engines/hfbi/dbent/mod.rs
@@ -89,8 +89,8 @@ mod dbent_private_tests {
     ) -> RecordHFBI {
         RecordHFBI {
             specie: SpecieHFBI {
-                nome_comune: "Test Specie",
-                codice_specie: "TS",
+                nome_comune: "Test Specie".to_string(),
+                codice_specie: "TS".to_string(),
                 autoctono: true,
                 gruppo_eco,
                 // The "..Default::default()" has been removed and all fields are now set explicitly