+++ src/engines/hfbi/dhzp/mod.rs
@@ -99,8 +99,8 @@ mod dhzp_private_tests {
     ) -> RecordHFBI {
         RecordHFBI {
             specie: SpecieHFBI {
-                nome_comune: "Test Specie",
-                codice_specie: "TS",
+                nome_comune: "Test Specie".to_string(),
+                codice_specie: "TS".to_string(),
                 autoctono: true,
                 gruppo_eco,
                 gruppo_trofico: GruppoTrofHFBI {