+++ src/engines/hfbi/dmig/mod.rs
@@ -101,8 +101,8 @@ mod dmig_private_tests {
     ) -> RecordHFBI {
         RecordHFBI {
             specie: SpecieHFBI {
-                nome_comune: "Test Specie",
-                codice_specie: codice_specie,
+                nome_comune: "Test Specie".to_string(),
+                codice_specie: codice_specie.to_string(),
                 autoctono: true,
                 gruppo_eco,
                 gruppo_trofico: GruppoTrofHFBI {